+++ Runtime/Enums/SingletonMode.cs
@@ -1,0 +1,8 @@
+﻿namespace TheRealIronDuck.Ducktion.Enums
+{
+    public enum SingletonMode
+    {
+        Singleton,
+        NonSingleton
+    }
+}