+++ Tests/Editor/DucktionTestConfig.cs
@@ -1,4 +1,5 @@
-﻿using TheRealIronDuck.Ducktion.Logging;
+﻿using TheRealIronDuck.Ducktion.Enums;
+using TheRealIronDuck.Ducktion.Logging;
 
 namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor
 {
@@ -6,9 +7,18 @@ namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor
     {
         public readonly bool CreateContainer;
         public readonly LogLevel LogLevel;
+        public readonly bool EnableAutoResolve;
+        public readonly SingletonMode AutoResolveSingletonMode;
 
-        public DucktionTestConfig(bool createContainer = true, LogLevel logLevel = LogLevel.Disabled)
+        public DucktionTestConfig(
+            bool createContainer = true,
+            LogLevel logLevel = LogLevel.Disabled,
+            bool enableAutoResolve = false,
+            SingletonMode autoResolveSingletonMode = SingletonMode.Singleton
+        )
         {
+            EnableAutoResolve = enableAutoResolve;
+            AutoResolveSingletonMode = autoResolveSingletonMode;
             CreateContainer = createContainer;
             LogLevel = logLevel;
         }